+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0 - 2024-04-05]
+## [1.1.0 - 2024-04-10]
 ### Added
 - Dockerfile
 - Single Spa
@@ -56,3 +56,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Bridgehead can accept only if not accepted and reject only if not rejected
 - Fetch labels of script, application form and votum
 - Update active bridgehead after clicking on bridgehead in bridgehead overview
+- Exist invited users -> can show bridgehead admin buttons